requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
joblib
sqlalchemy
pydantic-settings
geopy
httpx
